requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
EasyProcess==1.1
entrypoint2==1.1
Flask==2.2.2
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
lazyme==0.0.27
MarkupSafe==2.1.2
matplotlib==3.6.3
mediapipe==0.9.1.0
MouseInfo==0.1.3
mss==7.0.1
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
openai==0.26.4
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
pathy==0.10.1
Pillow==9.4.0
playsound==1.3.0
preshed==3.0.8
protobuf==3.20.3
psutil==5.9.4
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycaw==20220416
pycountry==22.3.5
pydantic==1.10.4
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pyttsx3==2.90
pytube==12.1.2
pytweening==1.0.4
pywhatkit==5.4
pywin32==305
requests==2.28.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.9.0
speedtest-cli==2.1.3
srsly==2.4.5
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==1.1.1
Werkzeug==2.2.2
wikipedia==1.4.0
yarl==1.8.2
zipp==3.12.1
sourcedefender==10.0.13